#9 Team optimization component finished.
Create a component that optimize the team assign and create from personnel users is finished. A structure was created to assign personnel to subpart in the most optimal way. The structure that allows post-calculation assignments to be made according to the determined subpart and determined manpower was produced. Also in pre-made pars, some bugs fixed.